Migrating to MVI Arch

+++ common/data-models/src/commonMain/kotlin/com/shabinder/common/DownloadRecord.kt
@@ -1,10 +1,10 @@
 package com.shabinder.common
 
 data class DownloadRecord(
-    var id:Int = 0,
+    var id:Long = 0,
     var type:String,
     var name:String,
     var link:String,
     var coverUrl:String,
-    var totalFiles:Int = 1,
+    var totalFiles:Long = 1,
 )
